requirements.txt
aiohttp==1.0.5
discord.py==0.16.12
youtube_dl==2017.9.2
discord==0.0.2
PyNaCl==1.0.1
tornado==4.5.2
pillow==4.3.0